progress on JRA
